Handle multiple commands in a single job correctly
When a query string contains multiple commands the server will send
multiple command completion messages, each resulting in PGresult. In
between the commands the connection can go back into a busy state. To
handle this the busy state should be checked after processing each
result, bailing out to main loop if the job is not yet completed.

If we don't do this PQgetResult will block until the next command is
processed, blocking processing of other jobs.